Se valida que las 4 transacciones poseen funcionamiento del check para cambio en modo de calculo

+++ TransferApp/Models/SimpleTransactionsModel.cs
@@ -21,7 +21,7 @@
         public DateTime? IssueDateCheck { get; set; }
         public string Justify_AgentName { get; set; } = string.Empty;
         public DateTime? Justify_DateError { get; set; }
-        public byte CalculationMode { get; set; }
+        public int CalculationMode { get; set; }
         public int? Status { get; set; }
 
         public DateTime? DateC { get; set; }
